Install Broker as a Service.
