+++ src/Host/Broker/Impl/Microsoft.R.Host.Broker.csproj
@@ -72,8 +72,12 @@
   <ItemGroup>
     <Reference Include="System" />
     <Reference Include="System.ComponentModel.Composition" />
+    <Reference Include="System.Configuration.Install" />
+    <Reference Include="System.Data" />
     <Reference Include="System.Management" />
     <Reference Include="System.Net.Http" />
+    <Reference Include="System.ServiceProcess" />
+    <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
     <Compile Include="..\..\..\GlobalAssemblyInfo.cs">
@@ -82,6 +86,18 @@
     <Compile Include="About\LoadController.cs" />
     <Compile Include="About\AboutController.cs" />
     <Compile Include="ApiErrorResult.cs" />
+    <Compile Include="RBrokerServiceInstaller.cs">
+      <SubType>Component</SubType>
+    </Compile>
+    <Compile Include="RBrokerServiceInstaller.Designer.cs">
+      <DependentUpon>RBrokerServiceInstaller.cs</DependentUpon>
+    </Compile>
+    <Compile Include="Startup\BrokerService.cs">
+      <SubType>Component</SubType>
+    </Compile>
+    <Compile Include="Startup\BrokerService.Designer.cs">
+      <DependentUpon>BrokerService.cs</DependentUpon>
+    </Compile>
     <Compile Include="Extensions\BasicSignInContextExtensions.cs" />
     <Compile Include="Extensions\ExceptionExtensions.cs" />
     <Compile Include="Interpreters\Interpreter.cs" />
@@ -114,6 +130,7 @@
     <Compile Include="Security\Policies.cs" />
     <Compile Include="Security\TlsConfiguration.cs" />
     <Compile Include="Sessions\SessionStateChangedEventArgs.cs" />
+    <Compile Include="Startup\CommonStartup.cs" />
     <Compile Include="Startup\StartupOptions.cs" />
     <Compile Include="Startup\Program.cs" />
     <Compile Include="Sessions\Session.cs" />
@@ -165,6 +182,9 @@
     </FilesToSign>
   </ItemGroup>
   <ItemGroup>
+    <EmbeddedResource Include="RBrokerServiceInstaller.resx">
+      <DependentUpon>RBrokerServiceInstaller.cs</DependentUpon>
+    </EmbeddedResource>
     <EmbeddedResource Include="Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
       <LastGenOutput>Resources.Designer.cs</LastGenOutput>
