+++ src/Host/Broker/Impl/RBrokerServiceInstaller.resx
@@ -117,43 +117,13 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  <data name="Error_ClientTimedOut" xml:space="preserve">
-    <value>Client took too long to respond.</value>
-  </data>
-  <data name="Error_DeleteUserProfileFailed" xml:space="preserve">
-    <value>Failed to delete user profile for user {0}\{1}, with Win32 error {2}.</value>
-  </data>
-  <data name="Error_UserLogonFailed" xml:space="preserve">
-    <value>Failed to logon user {0}\{1}, with Win32 error {2}.</value>
-  </data>
-  <data name="Error_UserProfileCreateFailed" xml:space="preserve">
-    <value>Failed to create user profile for user {0}\{1}, with Win32 error {2}.</value>
-  </data>
-  <data name="Error_UserProfileDirectoryWasNotFound" xml:space="preserve">
-    <value>User profile directory expected but not found for user {0}\{1}, Win32 error {0}</value>
-  </data>
-  <data name="Error_UserProfileServiceError" xml:space="preserve">
-    <value>User profile service failed with exception: {0}.</value>
-  </data>
-  <data name="Info_DeletedUserProfile" xml:space="preserve">
-    <value>User profile deleted for user {0}\{1}.</value>
-  </data>
-  <data name="Info_DeletingUserProfile" xml:space="preserve">
-    <value>Deleting user profile for user {0}\{1}.</value>
-  </data>
-  <data name="Info_UserProfileAlreadyExists" xml:space="preserve">
-    <value>User profile already exists for user  {0}\{1}.</value>
-  </data>
-  <data name="Info_UserProfileCreated" xml:space="preserve">
-    <value>Created user profile for user {0}\{1}.</value>
-  </data>
-  <data name="Info_UserProfileDirectoryFound" xml:space="preserve">
-    <value>User profile directory was found for user {0}\{1}, path: {2}</value>
-  </data>
-  <data name="Text_ServiceDescription" xml:space="preserve">
-    <value>Creates Windows User profile for users connecting via R Host Broker</value>
-  </data>
-  <data name="Text_ServiceDisplayName" xml:space="preserve">
-    <value>R User Profile Service</value>
-  </data>
+  <metadata name="serviceProcessInstaller1.TrayLocation" type="System.Drawing.Point, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a">
+    <value>17, 17</value>
+  </metadata>
+  <metadata name="serviceInstaller1.TrayLocation" type="System.Drawing.Point, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a">
+    <value>196, 17</value>
+  </metadata>
+  <metadata name="$this.TrayLargeIcon" type="System.Boolean, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089">
+    <value>False</value>
+  </metadata>
 </root>
